Use int for score in Rating data model

+++ src/codemodder/codetf.py
@@ -102,7 +102,7 @@ class Strategy(Enum):
 
 
 class Rating(BaseModel):
-    score: float
+    score: int
     description: Optional[str] = None
